[YS-531] refactor: 상수, 유틸 함수 파일 위치 수정
