+++ src/app/user/profile/ProfilePage.constants.ts
@@ -1,0 +1,5 @@
+export const MATCH_TYPE_MAP = {
+  ALL: '전체',
+  OFFLINE: '대면',
+  ONLINE: '비대면',
+} as const;
